09.01.2019-16:16

+++ src/main/java/converter/Converter.java
@@ -5,13 +5,13 @@ import entity.Animal;
 
 public class Converter {
 
-            public Animal animalDtoToAnimal(AnimalDto animalDto){
-                Animal animal = new Animal();
-                animal.setAnimalId(animalDto.getAnimalId());
-                animal.setAnimalName(animalDto.getAnimalName());
-                animal.setIdCategory(animalDto.getIdCategory());
-                animal.setAnimalPicture(animalDto.getAnimalPicture());
-                animal.setAnimalSound(animalDto.getAnimalSound());
+    public Animal animalDtoToAnimal(AnimalDto animalDto){
+        Animal animal = new Animal();
+        animal.setAnimalId(animalDto.getAnimalId());
+        animal.setAnimalName(animalDto.getAnimalName());
+        animal.setIdCategory(animalDto.getIdCategory());
+        animal.setAnimalPicture(animalDto.getAnimalPicture());
+        animal.setAnimalSound(animalDto.getAnimalSound());
 
                 /*try {
                     if (!animalDto.getAnimalPicture().isEmpty() && !animalDto.getAnimalSound().isEmpty() ) {
@@ -22,21 +22,21 @@ public class Converter {
                     System.out.println("error_field_upload");
                 }*/
 
-                return animal;
-            }
+        return animal;
+    }
 
-            public AnimalDto animalToAnimalDto(Animal animal){
+    public AnimalDto animalToAnimalDto(Animal animal){
 
-                AnimalDto animalDto = new AnimalDto();
+        AnimalDto animalDto = new AnimalDto();
 
-                animalDto.setAnimalId(animal.getAnimalId());
-                animalDto.setAnimalName(animal.getAnimalName());
-                animalDto.setIdCategory(animal.getIdCategory());
+        animalDto.setAnimalId(animal.getAnimalId());
+        animalDto.setAnimalName(animal.getAnimalName());
+        animalDto.setIdCategory((int) animal.getIdCategory());
 //                animalDto.setAnimalPicture(animal.getAnimalPicture());
 //                animalDto.setAnimalSound(animal.getAnimalSound());
 
 
-                return animalDto;
+        return animalDto;
 
 
     }
